:fire: add: eunm 파일 데릭토리 수정
