+++ src/main/java/com/example/DongNaeJoGak/domain/comment/entity/enums/CommentReportType.java
@@ -1,4 +1,4 @@
-package com.example.DongNaeJoGak.domain.comment.entity;
+package com.example.DongNaeJoGak.domain.comment.entity.enums;
 
 public enum CommentReportType {
     ABUSE, SEXUAL, SPAM, PRIVACY, DEFAMATION, OTHER
