--- src/main/java/com/example/DongNaeJoGak/domain/idea/entity/IdeaReactionType.java
@@ -1,5 +1,0 @@
-package com.example.DongNaeJoGak.domain.idea.entity;
-
-public enum IdeaReactionType {
-    LIKE, DISLIKE
-}
