--- src/main/java/com/example/DongNaeJoGak/domain/idea/entity/IdeaStatus.java
@@ -1,5 +1,0 @@
-package com.example.DongNaeJoGak.domain.idea.entity;
-
-public enum IdeaStatus {
-    VOTING, REVIEW, APPROVED
-}
