+++ src/main/java/com/example/DongNaeJoGak/domain/idea/entity/enums/IdeaStatus.java
@@ -1,0 +1,5 @@
+package com.example.DongNaeJoGak.domain.idea.entity.enums;
+
+public enum IdeaStatus {
+    VOTING, REVIEW, APPROVED
+}
