--- src/main/java/com/example/DongNaeJoGak/domain/member/entity/ProviderType.java
@@ -1,5 +1,0 @@
-package com.example.DongNaeJoGak.domain.member.entity;
-
-public enum ProviderType {
-    KAKAO, GOOGLE
-}
