+++ src/main/java/com/example/DongNaeJoGak/domain/member/entity/enums/ProviderType.java
@@ -1,0 +1,5 @@
+package com.example.DongNaeJoGak.domain.member.entity.enums;
+
+public enum ProviderType {
+    KAKAO, GOOGLE
+}
